+++ config/config_test.yaml
@@ -14,7 +14,7 @@ proc:
   # It can be either 'all' for processing all the epochs,
   # A list of two integers for processing a range of epoches
   # or a list with the epoche numbers to be processed
-  epoch_to_process: [180, 338] #[180, 185] #  "all" #
+  epoch_to_process: [180, 182] # [180, 338] #[180, 185] #  "all" #
 
   #- Switches to find and track matches
   do_matching: true
@@ -78,9 +78,9 @@ other:
 matching:
   resize: [-1]
   keypoint_threshold: 0.0001
-  max_keypoints: 8192 #4096 #  10240
+  max_keypoints: 1024 # 8192 #4096 #  10240
   superglue: "outdoor"
-  match_threshold: 0.10
+  match_threshold: 0.3 #0.10
   viz_matches: true
   fast_viz: true
   force_cpu: false
@@ -95,7 +95,7 @@ matching:
 tracking:
   resize: [-1]
   keypoint_threshold: 0.005
-  max_keypoints: 8192
+  max_keypoints: 1024 #8192
   superglue: "outdoor"
   match_threshold: 0.3
   viz_matches: false
@@ -112,7 +112,7 @@ tracking:
 metashape:
   # Processing parameters:
   optimize_cameras: true
-  build_dense: true
+  build_dense: false
 
   # List containing camera location. Note, that the order may differ from that of icepy, because Metashape imports the cameras based on alphabetic order of the images that they belong to (e.g., IMG_1289.jpg that belongs to p2 comes before IMG_2814.jpg that belongs to p1!)
   #TODO: fix camera ordering issue in Metashape.
