+++ src/icepy/classes/camera.py
