+++ src/icepy/classes/epoch.py
