+++ src/icepy/io/export2textfile.py
@@ -3,8 +3,8 @@ import logging
 
 from pathlib import Path
 
-from ..base_classes.features import Features
-from ..base_classes.images import ImageDS
+from ..classes.features import Features
+from ..classes.images import ImageDS
 
 """
 Export keypoints and points3d to file
