+++ src/icepy/sfm/geometry.py
@@ -25,7 +25,7 @@ SOFTWARE.
 import numpy as np
 import cv2
 
-from ..base_classes.camera import Camera
+from ..classes.camera import Camera
 
 
 def estimate_pose(kpts0, kpts1, K0, K1, thresh, conf=0.9999):
