+++ src/icepy/utils/initialization.py
@@ -33,10 +33,11 @@ from pathlib import Path
 from typing import List, Union
 from datetime import datetime
 
-from ..base_classes.camera import Camera
-from ..base_classes.point_cloud import PointCloud
-from ..base_classes.images import Image, ImageDS
-from ..base_classes.targets import Targets
+from ..classes.camera import Camera
+from ..classes.features import Features
+from ..classes.point_cloud import PointCloud
+from ..classes.images import Image, ImageDS
+from ..classes.targets import Targets
 
 # This file defines the dictionary cfg which includes the default parameters of the pipeline.
 # The dictionary is updated/extended at runtime with the parameters defined by the user in the input yaml config file
@@ -282,6 +283,8 @@ class Inizialization:
 
         for epoch in self.cfg.proc.epoch_to_process:
             features[epoch] = dict.fromkeys(self.cams)
+            for cam in self.cams:
+                features[epoch][cam] = Features()
 
         self.features = features
         return self.features
